==> 01_get_census_api_data_1990s.Rmd <==
---
title: "Census API Data"
author: "Stan Ferguson"
date: "1/5/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```
### Choosing the data  

##### First, take a look at all of the available US Census APIs
``` {r}
# Load libraries
library(censusapi)
library(tidyr)
library(stringr)
library(dplyr)
library(tidyverse)

# Force read the R environment file
readRenviron("~/.Renviron")

# Create a new dataframe with the full list of available census APIs
apis <- listCensusApis()

CENSUS_KEY="d1618423da2fa2032781c64c8f39311867ed3053"

View(apis)
```


``` {r}
# Create a new dataframe to store the metadata for the 2018 ACS 5-year dataset


acs_vars_nineties <- listCensusMetadata(
 name = "pep/int_charagegroups",
 vintage = "1990",
 type = "variables"
)

View(acs_vars_nineties)
```


``` {r}
write.csv(acs_vars_nineties, "data/demographics/1990s_metadata.csv", row.names = FALSE)
```


``` {r}
CENSUS_KEY="d1618423da2fa2032781c64c8f39311867ed3053"

acs_data <- getCensus(
        name = "pep/int_charagegroups",
        vintage = "1990",
        region = "county:*",
        regionin = "state:47",
        key = CENSUS_KEY,
        vars = c("YEAR,COUNTY,RACE_SEX,HISP,POP,AGEGRP")
)
```


``` {r}
head(acs_data)
```


```{r}
acs_data <- acs_data %>%
  mutate(RACE_SEX =
    case_when
    (
      RACE_SEX %in% c("01","02") ~ "white",
      RACE_SEX %in% c("03","04") ~ "black",
      RACE_SEX %in% c("05","06","07","08") ~ "other"
  )) %>%
  group_by(YEAR,COUNTY,RACE_SEX) %>%
  summarise(sum(POP))

head(acs_data)
```


``` {r}
county_codes <- read.csv("data/county_codes/county_codes.csv")

county_codes <- county_codes %>%
  rename(COUNTY = county)

county_codes$COUNTY <- county_codes$COUNTY %>%
  str_sub(3,5)

acs_data_nineties <- 
        inner_join(acs_data, county_codes, by = "COUNTY", copy = FALSE)

head(acs_data_nineties)
```


``` {r}
acs_data_nineties <- acs_data_nineties %>%
  ungroup(COUNTY) %>%
  rename(pop = `sum(POP)`) %>%
  rename(year = YEAR) %>%
  rename(race = RACE_SEX)%>%
  select(year,name,race,pop)


head(acs_data_nineties)



# Write it out to a csv
write.csv(acs_data_nineties, "data/demographics/acs_data_nineties.csv", row.names = FALSE)
```
