==> 02_get_census_api_data_2010s.Rmd <==
---
title: "Census API Data"
author: "Stan Ferguson"
date: "1/14/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```
### Choosing the data  

##### First, take a look at all of the available US Census APIs
``` {r}
# Load libraries
library(censusapi)
library(tidyr)
library(stringr)
library(dplyr)
library(tidyverse)

# Force read the R environment file
readRenviron("~/.Renviron")

CENSUS_KEY="d1618423da2fa2032781c64c8f39311867ed3053"

# Create a new dataframe with the full list of available census APIs
apis <- listCensusApis()

# Take a look at the available APIs
# View(apis)
```


``` {r}



acs_tens <- c("2011","2012","2013","2014","2015","2016","2017","2018")

acs_vars_list <- c()
```




``` {r}
for (val in acs_tens)
 {
    acs_vars <- listCensusMetadata(
    name = "acs/acs5/profile",
    vintage = val,
    type = "variables"
) %>% 
      pull(name)
    
    if (val == '2011'){
      common_vars <- acs_vars
    }
    else {
      common_vars <- intersect(acs_vars, common_vars)
    }
}
```


``` {r}
View(common_vars)
```


```{r}
acs_vars_tens <- data.frame()

for (val in acs_tens)
 {
    acs_vars_prime <- listCensusMetadata(
    name = "acs/acs5/profile",
    vintage = val,
    type = "variables") %>%
      mutate(year = val)

    acs_vars_tens <- rbind(acs_vars_prime,acs_vars_tens)
}
```


```{r}
View(acs_vars_tens)
    
```

```{r}
write.csv(acs_vars_tens,"data/demographics/acs_vars_tens.csv")
```


```{r}
# Create a list of the label_02 values we want to filter on

vals <- c("DP05_0037E",
          "DP05_0038E",
          "DP05_0039E",
          "DP05_0044E",
          "DP05_0052E",
          "DP05_0057E",
          "DP05_0058E",
          "DP05_0066E",
          "DP05_0065E"
)
```

```{r}
# Split the `label` field into their own fields and keep the original `label` field
acs_vars_tens <- acs_vars_tens %>%
     separate(
     label,
     c('label_01', 'label_02', 'label_03', 'label_04', 'label_05', 'label_06', 'label_07', 'label_08'),
     sep = "!!",
     remove = FALSE,
     convert = FALSE
     )


```
```{r}
View(acs_vars_tens)
```




```{r}
# Create a new dataframe for only the variables we want and drop the columns we won't need
target_acs_vars <- acs_vars_tens %>% 
        filter(name %in% vals)
```


```{r}
View(target_acs_vars)
# Did it work?
# View(target_acs_vars)
```



``` {r}
# Count the number of times '!!' shows up as a separator in the `label` field
#max(str_count(acs_vars$label, "!!"))

# Split the `label` field into their own fields and keep the original `label` field
acs_vars <- acs_vars_tens %>%
     separate(
     label,
     c('label_01', 'label_02', 'label_03', 'label_04', 'label_05', 'label_06', 'label_07', 'label_08'),
     sep = "!!",
     remove = FALSE,
     convert = FALSE
     )

View(acs_vars)

acs_vars <- acs_vars %>%
  arrange(name)


View(acs_vars)
```


##### Get the data for all counties in TN

Link to the state / county codes here:
``` {r}
# Turn the name column from our target_acs_vars table into a list
target_names <- target_acs_vars$name

# Check to see if it worked
# View(target_names)
```


``` {r}
acs_vars_tens_data <- data.frame()

for (val in acs_tens)
  {
    acs_data <- getCensus(
    name = "acs/acs5/profile",
    vintage = val,
    region = "county:*",
    regionin = "state:47",
    key = CENSUS_KEY,
    vars = target_names
)
    acs_vars_tens_data <- rbind(acs_data,acs_vars_tens_data)
}
```


```{r}
View(acs_vars_tens_data)
```


``` {r}
# Pivot the dataframe so there aren't 195 columns
acs_vars_tens_data <- pivot_longer(
        acs_data,
        cols = starts_with("DP"),
        names_to = "name",
        values_to = "value"
)
```

```{r}
View(acs_vars_tens_data)
```



``` {r}
# Join the dataframe to the metadata dataframe to get all of the label information
acs_data_tens_join <-
        inner_join(acs_vars_tens_data, target_acs_vars, by = "name", copy = FALSE)
```


```{r}
View(acs_data_tens_join)
```


```{r}
county_codes_tens <- read.csv("data/county_codes/county_codes.csv")
```

```{r}

county_codes_tens$county <- county_codes_tens$county %>%
  str_sub(3,5)


View(county_codes_tens)

```


```{r}

acs_data_tens_join <-
        inner_join(acs_data_tens_join, county_codes_tens, by = "county", copy = FALSE)

```

```{r}
View(acs_data_tens_join)
```



``` {r}
# Write it out to a csv
write.csv(acs_data_tens_join, "data/demographics/acs_5y_2010s_data.csv", row.names = FALSE)
```
